% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KMroutines.R
\name{KMSim}
\alias{KMSim}
\title{Function to simulate survival times from a Kaplan-Meier estimator
It simulates survival times either from the original Kaplan-Meier estimator
or it fits a PGW distribution using least squares based on nlsp points}
\usage{
KMSim(
  KM.obj = NULL,
  times = NULL,
  survprob = NULL,
  n = 100,
  seed = 1234,
  type.sim = NULL,
  nlsp = NULL
)
}
\arguments{
\item{KM.obj}{: Kaplan-Meier object (a surv object)}

\item{times}{: times used to construct the Kaplan-Meier estimator}

\item{survprob}{: survival probabilities for the Kaplan-Meier estimator}

\item{n}{: number of observations}

\item{seed}{: seed used in the simulation}

\item{type.sim}{: "KM" for using the Kaplan-Meier estimator directly or "PGW" for using the PGW distribution fitted via least squares using nlsp points}

\item{nlsp}{: number of points to be used in the least squares fit (equally spaced between min and max)}
}
\description{
Function to simulate survival times from a Kaplan-Meier estimator
It simulates survival times either from the original Kaplan-Meier estimator
or it fits a PGW distribution using least squares based on nlsp points
}
