% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KMroutines.R
\name{rpgw}
\alias{rpgw}
\title{Random Number Generation Function PGW distribution}
\usage{
rpgw(n, eta, nu, delta)
}
\arguments{
\item{n:}{sample size}

\item{eta:}{scale parameter}

\item{nu:}{shape parameter}

\item{delta:}{shape parameter}
}
\value{
vector of simulated values
}
\description{
Random Number Generation Function PGW distribution
}
