% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KMroutines.R
\name{spgw}
\alias{spgw}
\title{PGW Survival Function}
\usage{
spgw(t, eta, nu, delta, log.p = FALSE)
}
\arguments{
\item{t:}{positive parameter}

\item{eta:}{scale parameter}

\item{nu:}{shape parameter}

\item{delta:}{shape parameter}

\item{log.p:}{return log survival (TRUE or FALSE)}
}
\value{
PGW survival function
}
\description{
PGW Survival Function
}
